Generalize party predicates to work on either Party or PartyAssist
