+++ Assets/Scripts/Stats/Party/PartyBehaviourType.cs
@@ -1,0 +1,8 @@
+namespace Frankie.Stats
+{
+    public enum PartyBehaviourType
+    {
+        Party,
+        PartyAssist
+    }
+}
